+++ src/main/kotlin/game/model/player/TileBuilding.kt
@@ -13,7 +13,7 @@ class TileBuilding(
     val bld: Building,
     val level: Int,
     val desiredLevel: Int,
-    val capacity: Int,
+    val capacity: Double,
     val stockpileCapacity: Int,
     val stockpileMax: Int,
     val options: Int,
